[REFT] clean code
